fix/#23 : 폴더 구조 수정
