+++ app/components/Header.tsx
@@ -1,5 +1,3 @@
-const Header = () => {
+export default function Header() {
   return <header className="bg-[#020202] h-[100px]"></header>;
-};
-
-export default Header;
+}
