+++ app/landing/page.tsx
@@ -1,29 +1,27 @@
-
 import 'swiper/css';
 import NavBar from '../components/NavBar';
 import Footer from '../components/Footer';
-import Section1 from '../components/LandingPage/Section1';
-import Section2 from '../components/LandingPage/Section2';
-import Section3 from '../components/LandingPage/Section3';
-import Section4 from '../components/LandingPage/Section4';
-import Section5 from '../components/LandingPage/Section5';
-import Section6 from '../components/LandingPage/Section6';
-import Section7 from '../components/LandingPage/Section7';
-
+import Section1 from './components/Section1';
+import Section2 from './components/Section2';
+import Section3 from './components/Section3';
+import Section4 from './components/Section4';
+import Section5 from './components/Section5';
+import Section6 from './components/Section6';
+import Section7 from './components/Section7';
 
 export default function LandingPage() {
   return (
     <>
       <NavBar />
-      <div className='mb-[100px]'/>
-      <Section1/>
-      <Section2/>
-      <Section3/>
-      <Section4/>
-      <Section5/>
-      <Section6/>
-      <Section7/>
-      <Footer/>      
+      <div className="mb-[100px]" />
+      <Section1 />
+      <Section2 />
+      <Section3 />
+      <Section4 />
+      <Section5 />
+      <Section6 />
+      <Section7 />
+      <Footer />
     </>
   );
 }
