+++ app/landing/review.ts
@@ -1,34 +1,33 @@
 export interface ReviewData {
-    text: string;
-    name: string;
-    school: string;
-  }
-  
-  export const testimonials: ReviewData[] = [
-    {
-      text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
-      name: '정수아1',
-      school: '홍익대학교 조형대학',
-    },
-    {
-      text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
-      name: '정수아2',
-      school: '홍익대학교 조형대학',
-    },
-    {
-      text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
-      name: '정수아3',
-      school: '홍익대학교 조형대학',
-    },
-    {
-      text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
-      name: '정수아4',
-      school: '홍익대학교 조형대학',
-    },
-    {
-      text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
-      name: '정수아5',
-      school: '홍익대학교 조형대학',
-    },
-  ];
-  
+  text: string;
+  name: string;
+  school: string;
+}
+
+export const testimonials: ReviewData[] = [
+  {
+    text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
+    name: '정수아1',
+    school: '홍익대학교 조형대학',
+  },
+  {
+    text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
+    name: '정수아2',
+    school: '홍익대학교 조형대학',
+  },
+  {
+    text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
+    name: '정수아3',
+    school: '홍익대학교 조형대학',
+  },
+  {
+    text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
+    name: '정수아4',
+    school: '홍익대학교 조형대학',
+  },
+  {
+    text: '“이런 점이 좋았구요 또 이런 점도 좋았어요 대충 이런 리뷰 텍스트들을 간결하게 넣으면 좋을 것 같습니다 우선은 이렇게 더미를 채워둘게요 허허 근데 리뷰는 어떻게 받나요? 일단은 이렇게만 채워둘게요 하하 어떡하지 더 써야하나”',
+    name: '정수아5',
+    school: '홍익대학교 조형대학',
+  },
+];
